+++ srcjs/app.js
@@ -144,22 +144,53 @@ $(function() {
           message.content +
           "</div>" +
           "</div>" +
-          "</div>",
-        // Events
-        on: {
-          open: function(popover) {
-            console.log("Popover open");
-          },
-          opened: function(popover) {
-            console.log("Popover opened");
-          }
-        }
+          "</div>"
       });
       $('[data-popover = "' + index + '"]').on("click", function() {
         popover.open();
       });
     });
   });
+
+
+  // New popover API shinyMobile 2.0.0
+  Shiny.addCustomMessageHandler('add_popover', function(message) {
+    // We store all created instances in app data so that we don't
+    // recreate them later if they exist ...
+    if (app.data.popovers[message.targetEl] === undefined) {
+      // Only create if popover is enable for the current targetEl
+      if (!$(message.targetEl).hasClass('popover-disabled')) {
+        // popover HTML layout
+        message.content = `
+        <div class="popover">
+          <div class="popover-angle"></div>
+          <div class="popover-inner">
+            <div class="block">${message.content}</div>
+          </div>
+        </div>
+        `;
+
+        // create instance
+        var p = app.popover.create(message);
+        // Open popover
+        p.open();
+        // Storage in app data (popovers array)
+        app.data.popovers[message.targetEl] = p;
+        }
+    } else {
+      // Only show if popover is enable for the current targetEl
+      if (!$(message.targetEl).hasClass('popover-disabled')) {
+        app.data.popovers[message.targetEl].open();
+      }
+    }
+  });
+
+
+  Shiny.addCustomMessageHandler('toggle_popover', function(message) {
+    // Simply adds a class to indicate that the popover has to be hidden
+    $(message).toggleClass('popover-disabled');
+  });
+
 
   // handle toasts
   Shiny.addCustomMessageHandler("toast", function(message) {